// src/app/models/movclis.ts
export interface Movclis {
    idcli: number;
    fechamov: string;
    coa: string;
    concep1: number;
    poliza: string;
    tipag: string;
    bonificacion: number;
    recargo: number;
    importe: number;
    oper: number;
    ace: string;
    usuario: string;
    conse: number;
    feccap: string;
    idmvcli: number;
    idcobra: number;
    idusuario: number;
    concep: string;
    saldo: number;
}